java/src/test/java/com/studio3104/leetcode/explore/Learn/Recursion/I/RecurrenceRelation/ReverseLinkedList/SolutionTest.java
package com.studio3104.leetcode.explore.Learn.Recursion.I.RecurrenceRelation.ReverseLinkedList;

import com.studio3104.leetcode.structure.ListNode;
import com.studio3104.leetcode.structure.ListNodeConverter;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.converter.ConvertWith;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

class SolutionTest {
    Solution solution = new Solution();

    static Stream<Arguments> argumentsProvider() {
        return Stream.of(
                Arguments.of(new int[]{1, 2, 3, 4, 5}, new int[]{5, 4, 3, 2, 1})
        );
    }

    @ParameterizedTest
    @MethodSource("argumentsProvider")
    void reverseList(
            @ConvertWith(ListNodeConverter.class) ListNode head,
            @ConvertWith(ListNodeConverter.class) ListNode expected
    ) {
        ListNode actual = solution.reverseList(head);
        Assertions.assertEquals(expected, actual);
    }
}
